+++ flake.nix
@@ -3,7 +3,8 @@
 
   inputs = {
     nixpkgs = {
-      url = "github:nixos/nixpkgs/nixos-25.11";
+      # until crates.io fix fully backported
+      url = "github:nixos/nixpkgs?tag=25.11&rev=c767db50e209f33ffce3c18165b36101079d367d";
     };
     fenix = {
       url = "github:nix-community/fenix/monthly";